-Updated price in Product class to be a double instead of a float -Working on method to initialize ArrayList of products

+++ workbook3/SearchInventory/src/main/java/com/pluralsight/Product.java
@@ -5,10 +5,10 @@ public class Product {
     // Class variables
     private int id;
     private String name;
-    private float price;
+    private double price;
 
     // Product constructor
-    public Product(int id, String name, float price) {
+    public Product(int id, String name, double price) {
         this.id = id;
         this.name = name;
         this.price = price;
@@ -23,7 +23,7 @@ public class Product {
         return this.name;
     }
 
-    public float getPrice() {
+    public double getPrice() {
         return this.price;
     }
 }
